pyproject.toml
[project]
name = "dspy-react-machina"
version = "0.1.0"
description = "Alternative ReAct implementation for DSPy"
readme = "README.md"
authors = [
    { name = "Arthur Moura Carvalho", email = "armoucar@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.0",
]

[project.scripts]
tests = "scripts.test_runner:run_tests"
tests-coverage = "scripts.test_runner:run_tests_with_coverage"
quality-check = "scripts.quality_check:run_quality_checks"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.13.3",
    "pytest>=8.1.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "pyright>=1.1.0",
    "pre-commit>=4.0.0",
]
examples = [
    "arize-phoenix>=12.2.0",
    "openinference-instrumentation-litellm>=0.1.0",
    "openinference-instrumentation-dspy>=0.1.0",
    "opentelemetry-api>=1.0.0",
    "opentelemetry-exporter-otlp>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
]

[tool.ruff]
line-length = 120

[tool.pyright]
include = ["src"]
ignore = ["tests", "examples"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingTypeStubs = false

[tool.coverage.run]
omit = ["src/dspy_react_machina/_debug_utils.py"]
